+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Added
+
+- Add `tinty config` subcommand to return config related information
+
 ## [0.12.0] - 2024-04-29
 
 ### Added